Revamped JASP Tests
