+++ JASP-Tests/analyses/test_summarystatsttestbayesianonesample.R
