+++ JASP-Tests/core/main.cpp
@@ -20,23 +20,16 @@
 
 int main(int argc, char *argv[])
 {
-    QApplication a(argc, argv);
+	QApplication a(argc, argv);
+	int failures = AutomatedTests::run(argc, argv);
 
-    int failures = AutomatedTests::run(argc, argv);
+	qDebug() << "##########";
+	if (failures == 0) {
+		qDebug() << "ALL TESTS PASSED";
+	} else {
+		qDebug() << failures << " TESTS FAILED!";
+	}
+	qDebug() << "##########";
 
-    qDebug() << "--------------------------------------------------------------------------------------";
-
-    if (failures == 0)
-    {
-      qDebug() << "ALL TESTS PASSED";
-    }
-    else
-    {
-      qDebug() << failures << " TESTS FAILED!";
-    }
-
-    qDebug() << "--------------------------------------------------------------------------------------";
-
-    return 0;
+	return 0;
 }
-
