+++ JASP-Tests/core/textfileread_test.cpp
